Added support for option -p
Delete unnecessery files 'set_opt.c' and 'set_opt.h'.
All functions for support '-p' are on new file 'work_with_param.c'.
If argument of option '-p' is a number, then the number (as string) is writing on'dane.txt'. If argument is string, then it is printing on the screen.