Fixed bug: no reading from a file with one number
Changed function 'param_read_from_file()'.
The function goes to end of file. Next, on the loop 'do-while', the function read previous char and checks, it is correct char. If char is corrent, 'num_detection_flag' is set up. When the loop finds uncorrect char (e.g. space, ';' or '\n'), the start of the last number in the file has been detected. Next, the function copies the string of last number to str_nr[] and return the result.

Added checking, that chars are correct ('0'-'9', '-', '.').